+++ PufferyKit/Package.swift
@@ -31,6 +31,7 @@ let package = Package(
         .package(name: "KeychainSwift", url: "https://github.com/evgenyneu/keychain-swift.git", from: "19.0.0"),
 //        .package(name: "AckeeTracker", url: "https://github.com/vknabel/AckeeTracker-Swift.git", from: "0.1.0"),
         .package(url: "https://github.com/pointfreeco/swift-composable-architecture.git", from: "0.8.0"),
+        .package(url: "https://github.com/nachonavarro/Pages.git", from: "0.1.5"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -40,6 +41,7 @@ let package = Package(
             dependencies: [
                 "PufferyKit",
                 "RegistrationModule",
+                "GettingStartedModule",
                 .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
             ]
         ),
@@ -57,6 +59,14 @@ let package = Package(
                 .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
             ]
         ),
+        .target(
+            name: "GettingStartedModule",
+            dependencies: [
+                "Pages",
+                "Overture",
+                "PufferyKit",
+                .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
+        ]),
         .testTarget(
             name: "PufferyKitTests",
             dependencies: [