+++ PufferyKit/Sources/DesignSystem/SlideImage.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+
+public struct SlideImage: View {
+    var image: Image
+    
+    public init(_ name: String) {
+        image = Image(name)
+    }
+    
+    public init(image: Image) {
+        self.image = image
+    }
+    
+    public var body: some View {
+        image
+            .padding()
+    }
+}