+++ PufferyKit/Sources/RegistrationModule/RegistrationPage.swift
@@ -1,10 +1,3 @@
-//
-//  RegistrationPage.swift
-//
-//
-//  Created by Valentin Knabel on 03.05.20.
-//
-
 import ComposableArchitecture
 import SwiftUI
 import PlatformSupport
@@ -13,7 +6,7 @@ import PufferyKit
 
 public struct RegistrationPage: View {
     var onFinish: () -> Void
-
+    
     @ObservedObject private var keyboard = Keyboard()
     let store: ComposableArchitecture.Store<RegistrationState, RegistrationAction>
     
@@ -21,82 +14,68 @@ public struct RegistrationPage: View {
         self.onFinish = onFinish
         self.store = store
     }
-
+    
     public var body: some View {
         WithViewStore(self.store) { viewModel in
-            VStack {
-                Spacer()
+            VStack(alignment: .leading) {
+                Text("GettingStarted.Registration.PrivacyOptionalEmail")
+                    .lineLimit(nil)
+                    .fixedSize(horizontal: false, vertical: true)
 
-                VStack {
+                EmailTextField(onFinish: onFinish, store: store)
+                
+                Text("GettingStarted.Registration.EmailOptIn")
+                    .opacity(0.5)
+                
+                if viewModel.email.isEmpty || keyboard.isActive {
                     Button(action: { viewModel.send(.shouldRegister(onFinish: self.onFinish)) }) {
                         Text("GettingStarted.Registration.Anonymous")
                     }
-                    .buttonStyle(RoundedButtonStyle())
-                    .show(when: !self.keyboard.isActive)
-                    .transition(.opacity)
-                    .disabled(viewModel.activity.inProgress)
-
-                    RegistrationTerms()
+                        .buttonStyle(RoundedButtonStyle())
+                        .transition(.opacity)
+                        .disabled(viewModel.activity.inProgress)
+                } else {
+                    Button(action: { viewModel.send(.shouldRegister(onFinish: self.onFinish)) }) {
+                        Text("GettingStarted.Registration.WithEmail")
+                    }
+                        .buttonStyle(RoundedButtonStyle())
+                        .transition(.opacity)
+                        .disabled(viewModel.activity.inProgress || viewModel.email.isEmpty)
                 }
-                Spacer()
-
+                RegistrationTerms()
+                    .padding(.vertical)
+                                
                 HStack {
                     VStack { Divider() }
-                    Text("OR")
+                    Text("GettingStarted.Registration.RegistrationLoginDelimiter")
                         .italic()
                         .font(.headline)
                     VStack { Divider() }
-                }.show(when: !self.keyboard.isActive).transition(.opacity)
-
-                VStack {
-                    TextField("GettingStarted.Login.EmailPlaceholder", text: Binding(get: { viewModel.email }, set: { viewModel.send(.updateEmail($0)) }), onCommit: { viewModel.send(.shouldLogin(onFinish: self.onFinish)) })
-                        .textFieldStyle(RoundedTextFieldStyle())
-                        .multilineTextAlignment(.center)
-                        .keyboardType(.emailAddress)
-                        .textContentType(.emailAddress)
-                        .autocapitalization(.none)
-                        .disableAutocorrection(true)
-                        .disabled(viewModel.activity.inProgress)
-
-                    Button(action: { viewModel.send(.shouldLogin(onFinish: self.onFinish)) }) {
-                        Text("GettingStarted.Login.Perform")
-                    }
-                    .disabled(viewModel.email.isEmpty || viewModel.activity.inProgress)
-                    .sheet(isPresented: viewModel.binding(
-                        get: { $0.shouldCheckEmails },
-                        send: RegistrationAction.showCheckEmails
-                    )
-                    ) {
-                        EmailConfirmationPage(email: viewModel.email)
-                    }
-
-                    RegistrationTerms()
                 }
-                .padding()
-                .padding(.bottom, self.keyboard.currentHeight)
-                .animation(.default)
-
-                Spacer()
-            }.buttonStyle(RoundedButtonStyle())
-                .trackAppearence("registration", using: Current.tracker)
-                .alert(item: Binding.constant(viewModel.activity.failedError)) {
-                    Alert(
-                        title: Text("GettingStarted.Registration.Failed"),
-                        message: Text($0.localizedDescription)
-                    )
+                .show(when: !self.keyboard.isActive).transition(.opacity)
+                
+                NavigationLink(
+                    "GettingStarted.Registration.LogInWithEmail",
+                    destination: EmailLoginPage(onFinish: onFinish, store: store)
+                )
+                    .buttonStyle(RoundedButtonStyle())
+                    .accentColor(Color.gray)
+                    .opacity(0.5)
+                    .show(when: !self.keyboard.isActive)
+            }
+            .padding(.horizontal, 20)
+            .multilineTextAlignment(.leading)
+            .alert(item: Binding.constant(viewModel.activity.failedError)) {
+                Alert(
+                    title: Text("GettingStarted.Registration.Failed"),
+                    message: Text($0.localizedDescription)
+                )
+            }
+            .sheet(isPresented: Binding(get: { viewModel.showsWelcomePage }, set: { viewModel.send(.showWelcomePage($0)) })) {
+                NavigationView {
+                    WelcomePage()
                 }
+            }
         }
     }
 }
-
-#if DEBUG
-    struct RegistrationPage_Previews: PreviewProvider {
-        static var previews: some View {
-            RegistrationPage(onFinish: {}, store: ComposableArchitecture.Store<RegistrationState, RegistrationAction>(
-                initialState: RegistrationState(),
-                reducer: registrationReducer,
-                environment: RegistrationEnvironment.live()
-            ))
-        }
-    }
-#endif