+++ PufferyWidget/MessageEntry.swift
@@ -1,5 +1,5 @@
-import WidgetKit
 import PufferyKit
+import WidgetKit
 
 struct MessageEntry: TimelineEntry {
     let message: Message?
@@ -7,7 +7,7 @@ struct MessageEntry: TimelineEntry {
 
     var date: Date {
         if let message = message {
-            return Date(timeIntervalSinceNow: min(3600, -1 * message.createdAt.timeIntervalSinceNow))
+            return Date(timeIntervalSinceNow: min(3600, -1 * message.createdAt.timeIntervalSinceNow / 2))
         } else {
             return Date(timeIntervalSinceNow: 60)
         }